+++ src/components/application-layout.tsx
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2025 SecPal
+// SPDX-FileCopyrightText: 2026 SecPal
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import { useNavigate, useLocation } from "react-router-dom";
@@ -34,18 +34,6 @@ function HomeIcon(props: React.SVGProps<SVGSVGElement>) {
       <path
         fillRule="evenodd"
         d="M9.293 2.293a1 1 0 0 1 1.414 0l7 7A1 1 0 0 1 17 11h-1v6a1 1 0 0 1-1 1h-2a1 1 0 0 1-1-1v-3a1 1 0 0 0-1-1H9a1 1 0 0 0-1 1v3a1 1 0 0 1-1 1H5a1 1 0 0 1-1-1v-6H3a1 1 0 0 1-.707-1.707l7-7Z"
-        clipRule="evenodd"
-      />
-    </svg>
-  );
-}
-
-function KeyIcon(props: React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg data-slot="icon" viewBox="0 0 20 20" fill="currentColor" {...props}>
-      <path
-        fillRule="evenodd"
-        d="M8 7a5 5 0 1 1 3.61 4.804l-1.903 1.903A1 1 0 0 1 9 14H8v1a1 1 0 0 1-1 1H6v1a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1v-2a1 1 0 0 1 .293-.707L8.196 8.39A5.002 5.002 0 0 1 8 7Zm5-3a.75.75 0 0 0 0 1.5A1.5 1.5 0 0 1 14.5 7 .75.75 0 0 0 16 7a3 3 0 0 0-3-3Z"
         clipRule="evenodd"
       />
     </svg>
@@ -206,9 +194,6 @@ export function ApplicationLayout({ children }: { children: React.ReactNode }) {
             >
               <Trans>Home</Trans>
             </NavbarItem>
-            <NavbarItem href="/secrets" current={isCurrentPath("/secrets")}>
-              <Trans>Secrets</Trans>
-            </NavbarItem>
             {hasOrgAccess && (
               <NavbarItem
                 href="/organization"
@@ -278,12 +263,6 @@ export function ApplicationLayout({ children }: { children: React.ReactNode }) {
                 <HomeIcon />
                 <SidebarLabel>
                   <Trans>Home</Trans>
-                </SidebarLabel>
-              </SidebarItem>
-              <SidebarItem href="/secrets" current={isCurrentPath("/secrets")}>
-                <KeyIcon />
-                <SidebarLabel>
-                  <Trans>Secrets</Trans>
                 </SidebarLabel>
               </SidebarItem>
               {hasOrgAccess && (